pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "leancall"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lark>=1.3.1",
    "lean-interact>=0.10.5",
]


[tool.hatchling.build]
packages = ["src/leancall"]

[dependency-groups]
dev = ["hypothesis>=6.150.2", "numpy>=1.26", "pytest>=9.0.2"]
extra = [
    "cffi>=2.0.0",
    "numpy>=2.4.1",
]
